Move VolatileLayerClient online test to separate file
Move VolatileLayerClient online test to separate file.

Relates-to: OLPEDGE-720
