+++ olp-cpp-sdk-dataservice-write/tests/TestVolatileLayerClient.cpp
@@ -109,330 +109,6 @@ class VolatileLayerClientTestBase : public ::testing::TestWithParam<bool> {
   std::shared_ptr<std::vector<unsigned char>> data_;
 };
 
-class VolatileLayerClientOnlineTest : public VolatileLayerClientTestBase {
- protected:
-  static std::shared_ptr<olp::http::Network> s_network;
-
-  static void SetUpTestSuite() {
-    s_network = olp::client::OlpClientSettingsFactory::
-        CreateDefaultNetworkRequestHandler();
-  }
-
-  virtual std::shared_ptr<VolatileLayerClient> CreateVolatileLayerClient()
-      override {
-    auto network = s_network;
-
-    olp::authentication::Settings authentication_settings;
-    authentication_settings.token_endpoint_url =
-        CustomParameters::getArgument(kEndpoint);
-    authentication_settings.network_request_handler = network;
-
-    olp::authentication::TokenProviderDefault provider(
-        CustomParameters::getArgument(kAppid),
-        CustomParameters::getArgument(kSecret), authentication_settings);
-
-    olp::client::AuthenticationSettings auth_client_settings;
-    auth_client_settings.provider = provider;
-
-    olp::client::OlpClientSettings settings;
-    settings.authentication_settings = auth_client_settings;
-    settings.network_request_handler = network;
-
-    return std::make_shared<VolatileLayerClient>(
-        olp::client::HRN{GetTestCatalog()}, settings);
-  }
-};
-
-// Static network instance is necessary as it needs to outlive any created
-// clients. This is a known limitation as triggered send requests capture the
-// network instance inside the callbacks.
-std::shared_ptr<olp::http::Network> VolatileLayerClientOnlineTest::s_network;
-
-INSTANTIATE_TEST_SUITE_P(TestOnline, VolatileLayerClientOnlineTest,
-                         ::testing::Values(true));
-
-TEST_P(VolatileLayerClientOnlineTest, GetBaseVersionTest) {
-  auto volatileClient = CreateVolatileLayerClient();
-  auto response = volatileClient->GetBaseVersion().GetFuture().get();
-
-  ASSERT_TRUE(response.IsSuccessful());
-  auto versionResponse = response.GetResult();
-  ASSERT_GE(versionResponse.GetVersion(), 0);
-}
-
-TEST_P(VolatileLayerClientOnlineTest, StartBatchInvalidTest) {
-  auto volatileClient = CreateVolatileLayerClient();
-  auto response =
-      volatileClient->StartBatch(StartBatchRequest()).GetFuture().get();
-
-  ASSERT_FALSE(response.IsSuccessful());
-  ASSERT_FALSE(response.GetResult().GetId());
-  ASSERT_EQ(olp::client::ErrorCode::BadRequest,
-            response.GetError().GetErrorCode());
-
-  auto getBatchResponse =
-      volatileClient->GetBatch(response.GetResult()).GetFuture().get();
-
-  ASSERT_FALSE(getBatchResponse.IsSuccessful());
-
-  auto completeBatchResponse =
-      volatileClient->CompleteBatch(getBatchResponse.GetResult())
-          .GetFuture()
-          .get();
-  ASSERT_FALSE(completeBatchResponse.IsSuccessful());
-}
-
-TEST_P(VolatileLayerClientOnlineTest, StartBatchTest) {
-  auto volatileClient = CreateVolatileLayerClient();
-  auto response =
-      volatileClient
-          ->StartBatch(StartBatchRequest().WithLayers({GetTestLayer()}))
-          .GetFuture()
-          .get();
-
-  ASSERT_TRUE(response.IsSuccessful());
-  ASSERT_TRUE(response.GetResult().GetId());
-  ASSERT_NE("", response.GetResult().GetId().value());
-
-  auto getBatchResponse =
-      volatileClient->GetBatch(response.GetResult()).GetFuture().get();
-
-  ASSERT_TRUE(getBatchResponse.IsSuccessful());
-  ASSERT_EQ(response.GetResult().GetId().value(),
-            getBatchResponse.GetResult().GetId().value());
-  ASSERT_EQ("initialized",
-            getBatchResponse.GetResult().GetDetails()->GetState());
-
-  auto completeBatchResponse =
-      volatileClient->CompleteBatch(getBatchResponse.GetResult())
-          .GetFuture()
-          .get();
-  ASSERT_TRUE(completeBatchResponse.IsSuccessful());
-
-  for (int i = 0; i < 100; ++i) {
-    getBatchResponse =
-        volatileClient->GetBatch(response.GetResult()).GetFuture().get();
-
-    ASSERT_TRUE(getBatchResponse.IsSuccessful());
-    ASSERT_EQ(response.GetResult().GetId().value(),
-              getBatchResponse.GetResult().GetId().value());
-    if (getBatchResponse.GetResult().GetDetails()->GetState() != "succeeded") {
-      ASSERT_EQ("submitted",
-                getBatchResponse.GetResult().GetDetails()->GetState());
-    } else {
-      break;
-    }
-  }
-  // There are can be a case that GetBatch() is not "succeeded", but in
-  // "submitted" state even after 100 iterations,
-  // which actually means that there are might be a problem on the server side,
-  // (or just long delay). Thus, better to rewrite this test, or do not rely on
-  // the real server, but use mocked server.
-  // ASSERT_EQ("succeeded",
-  // getBatchResponse.GetResult().GetDetails()->GetState());
-}
-
-TEST_P(VolatileLayerClientOnlineTest, PublishToBatchTest) {
-  auto volatileClient = CreateVolatileLayerClient();
-  auto response =
-      volatileClient
-          ->StartBatch(StartBatchRequest().WithLayers({GetTestLayer()}))
-          .GetFuture()
-          .get();
-
-  ASSERT_TRUE(response.IsSuccessful());
-  ASSERT_TRUE(response.GetResult().GetId());
-  ASSERT_NE("", response.GetResult().GetId().value());
-
-  std::vector<PublishPartitionDataRequest> partition_requests;
-  PublishPartitionDataRequest partition_request;
-  partition_requests.push_back(
-      partition_request.WithLayerId(GetTestLayer()).WithPartitionId("123"));
-  partition_requests.push_back(partition_request.WithPartitionId("456"));
-
-  auto publishToBatchResponse =
-      volatileClient->PublishToBatch(response.GetResult(), partition_requests)
-          .GetFuture()
-          .get();
-  ASSERT_TRUE(publishToBatchResponse.IsSuccessful());
-
-  auto completeBatchResponse =
-      volatileClient->CompleteBatch(response.GetResult()).GetFuture().get();
-  ASSERT_TRUE(completeBatchResponse.IsSuccessful());
-
-  GetBatchResponse getBatchResponse;
-  for (int i = 0; i < 100; ++i) {
-    getBatchResponse =
-        volatileClient->GetBatch(response.GetResult()).GetFuture().get();
-
-    ASSERT_TRUE(getBatchResponse.IsSuccessful());
-    ASSERT_EQ(response.GetResult().GetId().value(),
-              getBatchResponse.GetResult().GetId().value());
-    if (getBatchResponse.GetResult().GetDetails()->GetState() != "succeeded") {
-      ASSERT_EQ("submitted",
-                getBatchResponse.GetResult().GetDetails()->GetState());
-    } else {
-      break;
-    }
-  }
-  // There are can be a case that GetBatch() is not "succeeded", but in
-  // "submitted" state even after 100 iterations,
-  // which actually means that there are might be a problem on the server side,
-  // (or just long delay). Thus, better to rewrite this test, or do not rely on
-  // the real server, but use mocked server.
-  // ASSERT_EQ("succeeded",
-  // getBatchResponse.GetResult().GetDetails()->GetState());
-}
-
-TEST_P(VolatileLayerClientOnlineTest, PublishToBatchInvalidTest) {
-  auto volatileClient = CreateVolatileLayerClient();
-  auto response =
-      volatileClient
-          ->StartBatch(StartBatchRequest().WithLayers({GetTestLayer()}))
-          .GetFuture()
-          .get();
-
-  ASSERT_TRUE(response.IsSuccessful());
-  ASSERT_TRUE(response.GetResult().GetId());
-  ASSERT_NE("", response.GetResult().GetId().value());
-
-  auto publishToBatchResponse =
-      volatileClient->PublishToBatch(response.GetResult(), {})
-          .GetFuture()
-          .get();
-  ASSERT_FALSE(publishToBatchResponse.IsSuccessful());
-
-  std::vector<PublishPartitionDataRequest> partition_requests{
-      PublishPartitionDataRequest{}, PublishPartitionDataRequest{}};
-
-  publishToBatchResponse =
-      volatileClient->PublishToBatch(response.GetResult(), partition_requests)
-          .GetFuture()
-          .get();
-  ASSERT_FALSE(publishToBatchResponse.IsSuccessful());
-
-  partition_requests.clear();
-  PublishPartitionDataRequest partition_request;
-  partition_requests.push_back(
-      partition_request.WithLayerId("foo").WithPartitionId("123"));
-  partition_requests.push_back(
-      partition_request.WithLayerId("bar").WithPartitionId("456"));
-
-  publishToBatchResponse =
-      volatileClient->PublishToBatch(response.GetResult(), partition_requests)
-          .GetFuture()
-          .get();
-  ASSERT_FALSE(publishToBatchResponse.IsSuccessful());
-}
-
-// Sometimes we receive the 500 internal server error,
-// thus looks loke the problem is on the server side.
-// Please, re-enable this test when switched to mocked server or
-// when the server will be more steady for testing
-TEST_P(VolatileLayerClientOnlineTest, DISABLED_StartBatchDeleteClientTest) {
-  auto volatileClient = CreateVolatileLayerClient();
-  auto response =
-      volatileClient
-          ->StartBatch(StartBatchRequest().WithLayers({GetTestLayer()}))
-          .GetFuture()
-          .get();
-
-  ASSERT_TRUE(response.IsSuccessful());
-  ASSERT_TRUE(response.GetResult().GetId());
-  ASSERT_NE("", response.GetResult().GetId().value());
-
-  auto getBatchFuture =
-      volatileClient->GetBatch(response.GetResult()).GetFuture();
-
-  volatileClient = nullptr;
-
-  auto getBatchResponse = getBatchFuture.get();
-  ASSERT_TRUE(getBatchResponse.IsSuccessful());
-  ASSERT_EQ(response.GetResult().GetId().value(),
-            getBatchResponse.GetResult().GetId().value());
-  ASSERT_EQ("initialized",
-            getBatchResponse.GetResult().GetDetails()->GetState());
-
-  volatileClient = CreateVolatileLayerClient();
-
-  auto completeBatchResponse =
-      volatileClient->CompleteBatch(getBatchResponse.GetResult())
-          .GetFuture()
-          .get();
-  ASSERT_TRUE(completeBatchResponse.IsSuccessful());
-
-  for (int i = 0; i < 100; ++i) {
-    getBatchResponse =
-        volatileClient->GetBatch(response.GetResult()).GetFuture().get();
-
-    ASSERT_TRUE(getBatchResponse.IsSuccessful());
-    ASSERT_EQ(response.GetResult().GetId().value(),
-              getBatchResponse.GetResult().GetId().value());
-    if (getBatchResponse.GetResult().GetDetails()->GetState() != "succeeded") {
-      ASSERT_EQ("submitted",
-                getBatchResponse.GetResult().GetDetails()->GetState());
-    } else {
-      break;
-    }
-  }
-  // There are can be a case that GetBatch() is not "succeeded", but in
-  // "submitted" state even after 100 iterations,
-  // which actually means that there are might be a problem on the server side,
-  // (or just long delay). Thus, better to rewrite this test, or do not rely on
-  // the real server, but use mocked server.
-  // ASSERT_EQ("succeeded",
-  // getBatchResponse.GetResult().GetDetails()->GetState());
-}
-
-TEST_P(VolatileLayerClientOnlineTest, cancellAllRequestsTest) {
-  auto volatileClient = CreateVolatileLayerClient();
-  auto future = volatileClient->GetBaseVersion().GetFuture();
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(10));
-  volatileClient->cancellAll();
-
-  auto response = future.get();
-  ASSERT_FALSE(response.IsSuccessful());
-  ASSERT_EQ(olp::client::ErrorCode::Cancelled,
-            response.GetError().GetErrorCode());
-}
-
-TEST_P(VolatileLayerClientOnlineTest, PublishData) {
-  auto response = client_
-                      ->PublishPartitionData(PublishPartitionDataRequest()
-                                                 .WithData(data_)
-                                                 .WithLayerId(GetTestLayer())
-                                                 .WithPartitionId("123"))
-                      .GetFuture()
-                      .get();
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-
-TEST_P(VolatileLayerClientOnlineTest, PublishDataAsync) {
-  std::promise<PublishPartitionDataResponse> response_promise;
-  bool call_is_async = true;
-
-  auto cancel_token = client_->PublishPartitionData(
-      PublishPartitionDataRequest()
-          .WithData(data_)
-          .WithLayerId(GetTestLayer())
-          .WithPartitionId("456"),
-      [&](const PublishPartitionDataResponse& response) {
-        call_is_async = false;
-        response_promise.set_value(response);
-      });
-
-  EXPECT_TRUE(call_is_async);
-  auto response_future = response_promise.get_future();
-  auto status = response_future.wait_for(std::chrono::seconds(30));
-  if (status != std::future_status::ready) {
-    cancel_token.cancel();
-  }
-  auto response = response_future.get();
-
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response));
-}
-
 #if 0
 olp::client::NetworkAsyncHandler volatileSetsPromiseWaitsAndReturns(
     std::shared_ptr<std::promise<void>> preSignal,
